Add back Nelder-Mead in ACORE
